hand-crafted anchor links are now properly handled
Before, a link that was intended to refer to an existing anchor in the
document, such as:

 [see the above section](#docker-vs-local)

was translated into:

 [see the above section](#see-the-above-section)

I added a goldmark parsing of the page so that we can retrieve the
headings and do a proper mapping between the Hugo anchors and the Devto
anchors. The above example should now look like this:

 [see the above section](#docker-vs-local)

Also, if an anchor isn't found in the document, you will be warned.